refactor: move search-related DOM manipulation into searchPageRenderer
